// src/lib/api/model/toolPart.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { ToolState } from './toolState'
import type { ToolPartMetadata } from './toolPartMetadata'

export interface ToolPart {
  id: string
  sessionID: string
  messageID: string
  type: 'tool'
  callID: string
  tool: string
  state: ToolState
  metadata?: ToolPartMetadata
}